refactor: BattleHeader 타이머 분리로 리렌더 최소화 (#258) - BattleTimer 컴포넌트로 타이머 로직 분리 - BattleHeader에서 타이머 상태/이펙트 제거 - 헤더 전체 리렌더 빈도 감소
